requirements.txt
kafka-python==2.0.2
pyspark==3.3.2
streamlit==1.20.0

